refactor: swagger docs packaging
