+++ src/main/kotlin/org/shangahi/sellio_backend/api/swagger/doc/CategoryDoc.kt
@@ -1,4 +1,4 @@
-package org.shangahi.sellio_backend.api.swagger
+package org.shangahi.sellio_backend.api.swagger.doc
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
@@ -9,6 +9,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse
 import org.shangahi.sellio_backend.api.dto.CategoryRequest
 import org.shangahi.sellio_backend.api.dto.response.CategoryResponse
 import org.shangahi.sellio_backend.api.dto.response.ErrorResponse
+import org.shangahi.sellio_backend.api.swagger.ErrorResponseExample
 
 annotation class CategoryDoc {
     @Operation(
